for #806: tk-shotgun-launchpublish refactor. (#5)
* for #806: tk-shotgun-launchpublish refactor.

The app works in a slightly different way now:
1. Given a Version or a PublishedFile, resolve which published file needs to be open.
Two default hook implementations are available.

2. Once a published file with a valid path has been found, try to launch it using a list of hooks
defined in the config. Each hook will be executed in order until one opens an application.
The default config will try to:

a. Open the published file with a predefined app (parameters app_path_windows, app_path_linux, app_path_mac)
If no predefined app is provided for the platform, or at all, or it failed to open the file, then:

b. Open the published file with a shotgun app. Look at the file extension, find a launch command/engine and launch
the Shotgun app.
If no command/engine is found, or if the engine failed to launch, then:

c. Open the published file with the default system app for that file.

It is very similar to what was previously achieved, except that before it used to discriminate the launch strategies a bit
differently, i.e.:

a. If the published file matches a viewer_extension, open with the viewer.
If no match, or if open with the viewer fails,

b. Open with a shotgun launch app/ engine.
If that fails, or if the extension does not match any known engines,

c. Open with the default system app.